requirements.txt
django-webpack-loader==0.3.3
